\docType{data}
\name{VignetteResults}
\alias{VignetteResults}
\title{
  Pre-computed results to use in the Vignette.
}
\description{
    This list contains pre-comuted time-intensite results used in the vignette.
}
\format{
  \describe{A list, containing 5 results shown in the vignette; refer to the text for details.
    \item{BC_DeltaSDNB_RobustResults}{}
    \item{BC_DeltaSDRM_RobustResults}{}
    \item{BC_OriginalResults}{}
    \item{LW_DeltaSD_RobustResults}{}
    \item{LW_DeltaSDD_RobustResults}{}
  }
}
